+++ Gruntfile.js
@@ -178,19 +178,6 @@ module.exports = function (grunt) {
 			}
 		},
 
-		// Jasmine test configuration
-		jasmine: {
-			src: config.js.files,
-			options: {
-				host: 'http://127.0.0.1:8000/',
-				specs: config.tests.src,
-				template: require('grunt-template-jasmine-requirejs'),
-				templateOptions: {
-					requireConfigFile: config.js.config
-				}
-			}
-		},
-
 		// Configuration for Karma test-runner
 		karma: {
 			options: {
@@ -236,20 +223,20 @@ module.exports = function (grunt) {
 
 			js: {
 				files: config.jsHintFiles,
-				tasks: ['jshint', 'jasmine']
+				tasks: ['jshint', 'connect:test', 'karma:unit']
 			},
 
 			karma: {
 				files: [config.jsHintFiles, config.tests.src],
-				tasks: ['connect:test', 'jasmine', 'karma:unit']
+				tasks: ['connect:test', 'karma:unit']
 			}
 		},
 
 		// Setup concurrent tasks
 		concurrent: {
 			deploy1: ['jshint', 'modernizr', 'sass:deploy', 'imagemin', 'copy'],
-			deploy2: ['requirejs', 'connect:test', 'jasmine'],
-			dev1: ['jshint', 'connect:test', 'jasmine', 'sass:dev', 'copy'],
+			deploy2: ['requirejs', 'connect:test', 'karma:unit'],
+			dev1: ['jshint', 'connect:test', 'karma:unit', 'sass:dev', 'copy'],
 			dev2: ['requirejs']
 		}
 	});
@@ -267,6 +254,6 @@ module.exports = function (grunt) {
 	grunt.registerTask('default', ['dev']);
 
 	// Travis CI task
-	grunt.registerTask('travis', ['jshint', 'connect:test', 'jasmine']);
+	grunt.registerTask('travis', ['jshint', 'connect:test', 'karma:unit']);
 
 };