+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### HEAD
 
+* Remove Jasmine task from Grunt
 * Add Karma as Test Suite
 * Add default logo.svg
 * Remove grunt-contrib-clean